src/requirements.txt
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
hiredis==2.2.3
idna==3.4
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-telegram-bot-calendar==1.0.5
PyYAML==6.0
redis==4.5.5
requests==2.31.0
six==1.16.0
telebot==0.0.5
urllib3==2.0.2
